requirements.txt
APScheduler==3.9.1
discord.py==1.7.3
python-decouple==3.6
tzlocal==4.2
